requirements.txt
django
djangorestframework
mysqlclient
gunicorn
django-heroku
whitenoise
